requirements.txt
playwright>=1.49.0
httpx>=0.27.0
PyYAML>=6.0.2
tenacity>=8.2.3
rich>=13.9.2
python-dotenv>=1.0.1
